# optimize binindex calculation