--- a/src/hist1d.jl
+++ b/src/hist1d.jl
@@ -12,7 +12,7 @@ function sample(h::Hist1D; n::Int=1)
 end
 
 @inline function _edge_binindex(r::AbstractRange, x::Real)
-    return floor(Int, (x - first(r)) / step(r)) + 1
+    return floor(Int, (x - first(r)) * inv(step(r))) + 1
 end
 
 @inline function _edge_binindex(r::AbstractRange{<:Integer}, x::Integer)