--- a/src/hist1d.jl
+++ b/src/hist1d.jl
@@ -12,7 +12,8 @@ function sample(h::Hist1D; n::Int=1)
 end
 
 @inline function _edge_binindex(r::AbstractRange, x::Real)
-    return floor(Int, (x - first(r)) * inv(step(r))) + 1
+    return Base.unsafe_trunc(Int, round((x - first(r)) * inv(step(r)), RoundDown)) + 1
+    # return floor(Int, (x - first(r)) * inv(step(r))) + 1 # uses safe trunc but fails on -Inf, Inf, NaN
 end
 
 @inline function _edge_binindex(r::AbstractRange{<:Integer}, x::Integer)