```diff
-    return floor(Int, (x - first(r)) / step(r)) + 1
+    return floor(Int, (x - first(r)) * inv(step(r))) + 1
```
apparently gives a ~20% speedup for histogram filling (3.1ms -> 2.5ms for randn(10^6)).

_______

If we also did 
```julia
return Base.unsafe_trunc(Int, round((x - first(r)) * inv(step(r)), RoundDown)) + 1
```
we can get another 20% (2.5ms -> 2.05ms)

`unsafe_trunc` avoids a `if $(Tf(typemin(Ti))) <= x < $(Tf(typemax(Ti)))` check. Currently we have
```julia
julia> Hist1D([-Inf,Inf,NaN], -1:0.5:1)
ERROR: InexactError: trunc(Int64, -Inf)
```
and if we used `unsafe_trunc`, we'd have
```julia
julia> Hist1D([-Inf,Inf,NaN], -1:0.5:1)
                ┌                              ┐
   [-1.0, -0.5) ┤ 0
   [-0.5,  0.0) ┤ 0
   [ 0.0,  0.5) ┤ 0
   [ 0.5,  1.0) ┤ 0
                └                              ┘
edges: -1.0:0.5:1.0
bin counts: [0, 0, 0, 0]
total count: 0
```