@Moelf 2021-08-31
Fine with the `unsafe_trunc` behavior, it's almost like free `skipmissing`........ But I wonder if it would contaminate bins in certain cases.