@aminnj 2021-08-31
The only downside is
### with `unsafe_trunc`
```julia
julia> Hist1D([-Inf,Inf,NaN], -1:0.5:1; overflow=true)
                ┌                              ┐
   [-1.0, -0.5) ┤▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 3
   [-0.5,  0.0) ┤ 0
   [ 0.0,  0.5) ┤ 0
   [ 0.5,  1.0) ┤ 0
                └                              ┘
edges: -1.0:0.5:1.0
bin counts: [3, 0, 0, 0]
total count: 3

### with regular `trunc`
```julia
julia> Hist1D([-Inf,Inf,NaN], -1:0.5:1; overflow=true)
ERROR: InexactError: trunc(Int64, -Inf)
```

So with default `overflow=false`, `unsafe_trunc` actually has better behavior. And with `overflow=true` they both are wrong, but at least `unsafe_trunc` doesn't crash :)