@Moelf 2021-08-31
I would argue in the case where it would go wrong, it's better to crash then return a bad histogram.