@aminnj 2021-08-31
OK, back to regular trunc. Speed is huge already.